requirements.txt
gunicorn
django_heroku
djangorestframework
django-cors-headers
django-cors-headers
django-secure